ANDROID: trusty: park threads when cpu is offline (v2)
If the thread is running when the cpu goes offline, it loses its
affinity and gets migrated to another cpu. This in turn can cause trusty
to stop responding as it might need to run on that specific cpu once it
is back online.

Park the thread while the cpu is offline to avoid this migration. The
kernel also has apis to re-bind the thread to a specific cpu each time
it is unparked, but none of those apis are exported to modules yet.

(v2): Also call kthread_set_per_cpu since kthread_create_on_cpu no
longer marks the created thread per-cpu

Bug: 266595872
Bug: 274202992
Change-Id: I6f311a830e47c227d404a8da312adbab5e0c1acf
Signed-off-by: Arve Hjønnevåg <arve@android.com>